+++ application/src/test-fixtures/java/org/opentripplanner/street/model/StreetModelForTest.java
@@ -57,9 +57,9 @@ public class StreetModelForTest {
     );
   }
 
-  public static StreetEdge streetEdge(Vertex vA, Vertex vB) {
+  public static StreetEdge streetEdge(StreetVertex vA, StreetVertex vB) {
     var meters = SphericalDistanceLibrary.distance(vA.getCoordinate(), vB.getCoordinate());
-    return streetEdge((StreetVertex) vA, (StreetVertex) vB, meters, StreetTraversalPermission.ALL);
+    return streetEdge(vA, vB, meters, StreetTraversalPermission.ALL);
   }
 
   public static StreetEdgeBuilder<?> streetEdgeBuilder(