+++ application/src/test/java/org/opentripplanner/graph_builder/module/linking/TestGraph.java
@@ -1,8 +1,8 @@
 package org.opentripplanner.graph_builder.module.linking;
 
-import org.opentripplanner.routing.linking.VertexLinker;
 import org.opentripplanner.routing.linking.VertexLinkerTestFactory;
 import org.opentripplanner.street.graph.Graph;
+import org.opentripplanner.street.linking.VertexLinker;
 import org.opentripplanner.street.model.edge.StreetTransitStopLink;
 import org.opentripplanner.street.model.vertex.TransitStopVertex;
 import org.opentripplanner.street.model.vertex.Vertex;