+++ street/src/test/java/org/opentripplanner/street/linking/PlatformLinkingTest.java
@@ -1,11 +1,11 @@
-package org.opentripplanner.routing.linking;
+package org.opentripplanner.street.linking;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.opentripplanner.routing.linking.TransitStopVertexBuilderFactory.ofStop;
-import static org.opentripplanner.routing.linking.VisibilityMode.COMPUTE_AREA_VISIBILITY_LINES;
+import static org.opentripplanner.street.linking.VisibilityMode.COMPUTE_AREA_VISIBILITY_LINES;
+import static org.opentripplanner.street.model.StreetModelFactory.intersectionVertex;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -17,9 +17,10 @@ import org.locationtech.jts.geom.LineString;
 import org.locationtech.jts.geom.Polygon;
 import org.opentripplanner.core.model.i18n.I18NString;
 import org.opentripplanner.core.model.i18n.LocalizedString;
-import org.opentripplanner.routing.graph.GraphDataFetcher;
 import org.opentripplanner.street.geometry.GeometryUtils;
 import org.opentripplanner.street.graph.Graph;
+import org.opentripplanner.street.graph.GraphDataFetcher;
+import org.opentripplanner.street.model.StreetModelFactory;
 import org.opentripplanner.street.model.StreetTraversalPermission;
 import org.opentripplanner.street.model.edge.Area;
 import org.opentripplanner.street.model.edge.AreaEdge;
@@ -33,7 +34,6 @@ import org.opentripplanner.street.model.vertex.TransitStopVertex;
 import org.opentripplanner.street.model.vertex.Vertex;
 import org.opentripplanner.street.search.TraverseMode;
 import org.opentripplanner.street.search.TraverseModeSet;
-import org.opentripplanner.streetadapter.VertexFactory;
 
 public class PlatformLinkingTest {
 
@@ -165,9 +165,15 @@ public class PlatformLinkingTest {
     }
     assertNotNull(ag);
 
-    var vertexFactory = new VertexFactory(graph.graph());
-    var v = vertexFactory.intersection("boardingLocation", 10.00000001, 60.00000001);
-    VertexLinkerTestFactory.of(graph.graph()).addPermanentAreaVertex(v, ag);
+    var v = intersectionVertex("boardingLocation", 10.00000001, 60.00000001);
+
+    var linker = new VertexLinker(
+      graph.graph(),
+      VisibilityMode.COMPUTE_AREA_VISIBILITY_LINES,
+      50,
+      true
+    );
+    linker.addPermanentAreaVertex(v, ag);
 
     // vertex links to the single visibility point with 2 edges
     assertEquals(10, graph.listEdges().size());
@@ -392,14 +398,14 @@ public class PlatformLinkingTest {
 
   private GraphDataFetcher prepareTest(Coordinate[] platform, int[] visible, Coordinate[] stops) {
     var graph = new Graph();
-    var vertexFactory = new VertexFactory(graph);
 
     ArrayList<IntersectionVertex> vertices = new ArrayList<>();
     Coordinate[] closedGeom = new Coordinate[platform.length + 1];
 
     for (int i = 0; i < platform.length; i++) {
       Coordinate c = platform[i];
-      var vertex = vertexFactory.intersection(String.valueOf(i), c.x, c.y);
+      var vertex = intersectionVertex(String.valueOf(i), c.y, c.x);
+      graph.addVertex(vertex);
       vertices.add(vertex);
       closedGeom[i] = c;
     }
@@ -449,7 +455,8 @@ public class PlatformLinkingTest {
 
     for (int i = 0; i < stops.length; i++) {
       Coordinate stop = stops[i];
-      vertexFactory.transitStop(ofStop(i, stop));
+      var stopVertex = StreetModelFactory.transitStopVertex(i, stop);
+      graph.addVertex(stopVertex);
     }
 
     graph.index();
@@ -458,7 +465,12 @@ public class PlatformLinkingTest {
   }
 
   private void linkStops(GraphDataFetcher graph, int maxAreaNodes, boolean permanent) {
-    var linker = new VertexLinker(graph.graph(), COMPUTE_AREA_VISIBILITY_LINES, maxAreaNodes);
+    var linker = new VertexLinker(
+      graph.graph(),
+      COMPUTE_AREA_VISIBILITY_LINES,
+      maxAreaNodes,
+      false
+    );
     for (TransitStopVertex tStop : graph.listStopVertices()) {
       if (permanent) {
         linker.linkVertexBidirectionallyPermanently(