+++ street/src/test/java/org/opentripplanner/street/linking/ScopedLinkingTest.java
@@ -1,9 +1,9 @@
-package org.opentripplanner.routing.linking;
+package org.opentripplanner.street.linking;
 
 import static com.google.common.truth.Truth.assertThat;
 
 import org.junit.jupiter.api.Test;
-import org.opentripplanner.street.model.StreetModelForTest;
+import org.opentripplanner.street.model.StreetModelFactory;
 import org.opentripplanner.street.model.vertex.IntersectionVertex;
 import org.opentripplanner.street.search.TraverseModeSet;
 
@@ -13,7 +13,7 @@ import org.opentripplanner.street.search.TraverseModeSet;
  */
 class ScopedLinkingTest {
 
-  private static final IntersectionVertex SPLIT = StreetModelForTest.intersectionVertex(0.05, 0.05);
+  private static final IntersectionVertex SPLIT = StreetModelFactory.intersectionVertex(0.05, 0.05);
 
   @Test
   void splitRequestScoped() {
@@ -51,10 +51,10 @@ class ScopedLinkingTest {
   }
 
   private static LinkingEnvironment buildEnv() {
-    var v1 = StreetModelForTest.intersectionVertex(0.0, 0.0);
-    var v2 = StreetModelForTest.intersectionVertex(0.1, 0.1);
+    var v1 = StreetModelFactory.intersectionVertex(0.0, 0.0);
+    var v2 = StreetModelFactory.intersectionVertex(0.1, 0.1);
 
-    StreetModelForTest.streetEdge(v1, v2);
+    StreetModelFactory.streetEdge(v1, v2);
 
     return new LinkingEnvironment(v1, v2);
   }