+++ street/src/test/java/org/opentripplanner/street/linking/TemporaryVerticesContainerTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.linking;
+package org.opentripplanner.street.linking;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.locationtech.jts.geom.Coordinate;
 import org.opentripplanner.core.model.i18n.NonLocalizedString;
 import org.opentripplanner.street.graph.Graph;
-import org.opentripplanner.street.model.StreetModelForTest;
+import org.opentripplanner.street.model.StreetModelFactory;
 import org.opentripplanner.street.model.StreetTraversalPermission;
 import org.opentripplanner.street.model.edge.TemporaryFreeEdge;
 import org.opentripplanner.street.model.vertex.TemporaryStreetLocation;
@@ -19,15 +19,15 @@ public class TemporaryVerticesContainerTest {
   @Test
   void temporaryChangesRemovedOnClose() {
     var graph = new Graph();
-    var intersectionVertexA = StreetModelForTest.intersectionVertex(new Coordinate(0, 0));
-    var intersectionVertexB = StreetModelForTest.intersectionVertex(new Coordinate(0, 1));
-    var intersectionVertexC = StreetModelForTest.intersectionVertex(new Coordinate(1, 0));
-    StreetModelForTest.streetEdge(
+    var intersectionVertexA = StreetModelFactory.intersectionVertex(new Coordinate(0, 0));
+    var intersectionVertexB = StreetModelFactory.intersectionVertex(new Coordinate(0, 1));
+    var intersectionVertexC = StreetModelFactory.intersectionVertex(new Coordinate(1, 0));
+    StreetModelFactory.streetEdge(
       intersectionVertexA,
       intersectionVertexB,
       StreetTraversalPermission.PEDESTRIAN
     );
-    var secondEdge = StreetModelForTest.streetEdge(
+    var secondEdge = StreetModelFactory.streetEdge(
       intersectionVertexB,
       intersectionVertexC,
       StreetTraversalPermission.PEDESTRIAN
@@ -37,7 +37,12 @@ public class TemporaryVerticesContainerTest {
     graph.addVertex(intersectionVertexC);
     graph.index();
     try (var container = new TemporaryVerticesContainer()) {
-      var vertexLinker = VertexLinkerTestFactory.of(graph);
+      var vertexLinker = new VertexLinker(
+        graph,
+        VisibilityMode.COMPUTE_AREA_VISIBILITY_LINES,
+        50,
+        true
+      );
       var temporaryLocation = new TemporaryStreetLocation(
         new Coordinate(0.5, 0.5),
         new NonLocalizedString("Temp location")