+++ doc/user/Changelog.md
@@ -101,6 +101,7 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Add running days to flex routing [#7281](https://github.com/opentripplanner/OpenTripPlanner/pull/7281)
 - Update Entur-specific emission data docs url in Transmodel custom documentation  [#7352](https://github.com/opentripplanner/OpenTripPlanner/pull/7352)
 - Fix booking time calculation for multi-day flex trip searches [#7255](https://github.com/opentripplanner/OpenTripPlanner/pull/7255)
+- Fix empty string handling for `banned.routes` and `banned.agencies` in GTFS API [#7340](https://github.com/opentripplanner/OpenTripPlanner/pull/7340)
 [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)
 
 ## 2.8.1 (2025-09-22)